# [WIP] Encode and decode structs lazily