… envelope/envelope_test.go
… gen/internal/tests/collision/collision.go
… gen/internal/tests/containers/containers.go
… gen/internal/tests/enum_conflict/enum_conflict.go
… gen/internal/tests/enums/enums.go
… gen/internal/tests/exceptions/exceptions.go
… gen/internal/tests/hyphenated-file/hyphenated-file.go
… gen/internal/tests/hyphenated_file/hyphenated_file.go
… gen/internal/tests/non_hyphenated/non_hyphenated.go
… gen/internal/tests/nozap/nozap.go
… gen/internal/tests/services/services.go
… gen/internal/tests/set_to_slice/set_to_slice.go
… gen/internal/tests/structs/structs.go
… gen/internal/tests/typedefs/typedefs.go
… gen/internal/tests/unions/unions.go
… gen/internal/tests/uuid_conflict/uuid_conflict.go
… gen/roundtrip_test.go
… gen/service_test.go
… internal/envelope/server_test.go
… protocol/binary/lazy_field_list.go
… protocol/binary/reader.go
… protocol/binary/writer.go
… protocol/binary_test.go
… wire/evaluate.go
… wire/lazy_field_list.go
… wire/value.go
… wire/value_equals.go