@kanads found out that some services were spending a lot of time
allocating and growing slices of `[]wire.Field` during deserialization.
Specifically, [on this line][1], where we append to the slice, we spend
a fair bit of CPU growing the slice.

    fields = append(fields, wire.Field{ID: fid, Value: val})

@kanads is experimenting with pre-allocating these slices to some sane
value in production. That's a good short-term fix, but the long-term
fix is likely to address this TODO in that function:

    // TODO(abg) add a lazy FieldList type instead of []Field.

[1]: https://github.com/thriftrw/thriftrw-go/blob/9c16419a8cced33850d40e7620e5dfe96e7b1a4f/protocol/binary/reader.go#L304

This change adds a lazy wire.FieldList type that behaves similarly to
the wire.ValueList and wire.MapItemList types.

    type FieldList interface {
        ForEach(func(Field) error) error
        Close()
    }

Using this, during deserialization we can return an implementation that
doesn't deserialize from the struct until ForEach is called.

---

I'm not completely sold on this yet because of nested structs: for
nested structs, we'll end up skipping over them several times. However,
this might still end up being more efficient than the many allocations
we make. Benchmarks pending.

```
name                                        old time/op    new time/op    delta
RoundTrip/PrimitiveOptionalStruct/Encode-4    3.14µs ±22%    2.21µs ±12%   -29.72%  (p=0.008 n=5+5)
RoundTrip/PrimitiveOptionalStruct/Decode-4    4.79µs ±17%    3.38µs ±19%   -29.49%  (p=0.016 n=5+5)
RoundTrip/Graph/Encode-4                      7.16µs ±23%    3.37µs ±28%   -52.92%  (p=0.008 n=5+5)
RoundTrip/Graph/Decode-4                      9.59µs ±16%    8.64µs ±11%      ~     (p=0.222 n=5+5)
RoundTrip/ContainersOfContainers/Encode-4     34.5µs ±11%    27.9µs ±10%   -19.11%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Decode-4     80.6µs ±16%    78.6µs ±33%      ~     (p=0.548 n=5+5)

name                                        old alloc/op   new alloc/op   delta
RoundTrip/PrimitiveOptionalStruct/Encode-4      704B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
RoundTrip/PrimitiveOptionalStruct/Decode-4    1.40kB ± 0%    0.09kB ± 0%   -93.71%  (p=0.008 n=5+5)
RoundTrip/Graph/Encode-4                      1.70kB ± 0%    0.02kB ± 0%   -98.59%  (p=0.008 n=5+5)
RoundTrip/Graph/Decode-4                      2.78kB ± 0%    0.56kB ± 0%   -79.72%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Encode-4     1.30kB ± 0%    0.41kB ± 0%   -68.71%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Decode-4     12.3kB ± 0%    11.0kB ± 0%   -10.66%  (p=0.008 n=5+5)

name                                        old allocs/op  new allocs/op  delta
RoundTrip/PrimitiveOptionalStruct/Encode-4      1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
RoundTrip/PrimitiveOptionalStruct/Decode-4      14.0 ± 0%      11.0 ± 0%   -21.43%  (p=0.008 n=5+5)
RoundTrip/Graph/Encode-4                        11.0 ± 0%       1.0 ± 0%   -90.91%  (p=0.008 n=5+5)
RoundTrip/Graph/Decode-4                        32.0 ± 0%      42.0 ± 0%   +31.25%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Encode-4       18.0 ± 0%      17.0 ± 0%    -5.56%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Decode-4        164 ± 0%       161 ± 0%    -1.59%  (p=0.008 n=5+5)
```