@abhinav 2022-06-16
No longer needed with streaming APIs.